Add a status filter to view notices page (#2530)
https://eaflood.atlassian.net/browse/WATER-5307

The main driver for this was our support for returned letters from Notify. When Notify hits our callback endpoint, we mark the notification as `RETURNED`.

Because a returned letter can happen weeks after the letter has been sent, and users can be sending numerous notices, the notice with the returned notification can be pushed back in the results.

This means that unless they know which notification has been returned, they would have to manually scan the results to find notices that have returned notifications.

The solution is to add a filter that allows them to find notices with matching notifications. For example, a user may wish to find all notices with 'returned' and 'errored' notifications. The filter will allow them to see all matches.

Thanks to [Record notice status rather than calculate](https://github.com/DEFRA/water-abstraction-system/pull/2506), we can do this by just looking at the counts in `status_counts`.