Update view licence comms page to show all (#2535)
[> Link goes here](https://eaflood.atlassian.net/browse/WATER-5339)

In support of dynamic due dates, we have implemented the ability to capture when a letter is returned via a [callback from GOV.UK Notify](https://docs.notifications.service.gov.uk/rest-api.html#returned-letters). (See [WATER-5296](https://eaflood.atlassian.net/browse/WATER-5296))

This means we'll know which notification letters were returned, and when we are notified of this. We'll be changing the status of the notification to `RETURNED`, and users will be able to find them in the view notices page.

We've also realised there were lots of notifications in our service which had been cancelled via the **Notify** dashboard, but we were still showing them as `SENT`. So, [we now handle getting a 'cancelled' status from Notify](https://github.com/DEFRA/water-abstraction-system/pull/2514).

Being clear and correct about notifications is essential because dynamic due dates mean we'll be setting the due date of our return logs based on when we can confirm the notification was sent.

So, it's crucial that our users can see _all_ the notifications we've attempted to send to customers, in case they have any queries about what has been determined.

This change does two things.

- It amends the query behind the view licence communications page to fetch _all_ notifications for a licence
- It updates the way the notifications are displayed to be consistent with all the changes we've made to the view notices, notice, and notification pages