Remove page x of y from view notice page title (#2537)
https://eaflood.atlassian.net/browse/WATER-4992

We originally built the page to dynamically include, for example, `(page 1 of 3)` when the number of notifications forced pagination to kick in.

However, we don't do this in bill runs, and have recently removed it from the view notices page. The only other place is the review licences page as part of two-part tariff bill runs.

So, before this ships, we remove this from the title so the page is consistent with the rest of the service.