Enhance error details in view notifications page
https://eaflood.atlassian.net/browse/WATER-5327

In [Update view notification to support all statuses](https://github.com/DEFRA/water-abstraction-system/pull/2507) we made changes that meant rather than just showing the details for `SENT` notifications, it could now show details for _all_ notifications.

Part of this was to provide a means for users to see why a notification had been marked as failed.

When we check the status of a notification with [GOV.UK Notify](https://www.notifications.service.gov.uk/) it will return one of a series of status values, some of which means the notification is in progress, some that it sent successfully, and others that tell us it failed.

For these we show that status, and the description they provide in their [documentation](https://docs.notifications.service.gov.uk/rest-api.html#email-status-descriptions).

For the rest, we just display a generic error message. But what we have realised is that there is useful feedback the user isn't getting. For example, when we try to send a letter, if Notify deems the address to be invalid it will reject the request. We have that reason saved against the notification. But because it never got to Notify we just display 'Internal system error'.

This change is about enhancing the details we display to a user for errored notifications. We've done further analysis in the records and looked at what error messages we have to determine when we can provide a more informed error description.