Fix view notice licence ref filter validation
https://eaflood.atlassian.net/browse/WATER-4992

Whilst testing the filters on our new view notice page we found that the validation for licence ref was complaining that the length of a real licence reference was too long.

We must have set it to 11 in a fever dream because we can clearly see we have licence references up to 20 characters in length!

This change ups the limit to 25, to give us some wiggle room.