Amend Notify description
validation-failure is specific to uploaded letters and we are confident we keep our content inside the printable area.

In our testing we get this message when we upload an address which is missing a postcode or country, or with a known invalid address. So we wanted to ensure that was included in the message so users don't overlook checking the address used.