Include addresses with special char in invalid rpt
https://eaflood.atlassian.net/browse/WATER-5314

When working to [Enhance error details in view notifications page](https://github.com/DEFRA/water-abstraction-system/pull/2538) we spotted the following error message in the logs.

```Address lines must not start with any of the following characters: @ ( ) = [ ] ” \\\\ / , < >```

We realised that [Notify](https://www.notifications.service.gov.uk/) is rejecting some of our notifications because the addresses have lines that start with these special characters.

We actually validate for them when users enter addresses manually setting up an ad-hoc notice. But we didn't put the two together when we compiled the invalid addresses report. We need to include those with address lines that start with any of these special characters (we checked, there are some!)