Migrate search to system (#2477)
https://eaflood.atlassian.net/browse/WATER-5304

We are migrating the internal search page (the root page of the internal service) from the legacy apps to our new `water-abstraction-system app`.

The existing page allows users to search using different types of values, for example, licence references, licence names, and return references.

We are implementing these different types of search into the new `system` page one at a time to manage the complexity.

This change is to implement searching for licences.

The display of the search results is currently very basic - the actual display format is being worked on by our service designer and will be adjusted to suit whatever design they come up with in future changes.

I haven't added a feature flag for this feature yet as we can't really use it as a replacement for the existing search until we've got other things we can search on. For the moment, this search page can just be accessed at `/system/search`.