Update how unit tests work
We got our first flakey fail in CI because the number of results did not match 5 exactly.

Should have spotted earlier, but we now try to avoid explicit comparisions of length or content when test the results of fetching from the DB because we don't know what other tests have added and left behind.