Change notices status filter to OR instead of AND (#2544)
https://eaflood.atlassian.net/browse/WATER-5307

The original belief was that the status filter on the view notices page would be used to find notices containing one or more notifications with the selected statuses.

For example, if I select 'Cancelled' and 'Returned', I should only expect to see notices that have both a cancelled AND returned notification.

However, in testing, it became clear that the expectation was that notices with either a cancelled OR a returned notification would be in the results.

So, this change amends the logic for the status filter to use `OR`, not `AND`.