Hide legacy paper forms when ad-hoc enabled
https://eaflood.atlassian.net/browse/WATER-5345

Ad-hoc is not itself a notice type, but a new setup 'journey' to use when creating a new notice. Specifically, it requires users to tell us the licence we are creating an ad-hoc notice for.

Since you need to provide a licence, our replacement for the legacy paper returns journey is part of the ad-hoc journey, rather than a separate link on the **Manage** page.

You can see why this gets a bit confusing!

When ad-hoc is enabled and our version of paper returns is live, we can't allow users to create paper returns using the legacy journey.