Prevent sending notice when all invalid
We only just spotted that if all the recipients turn out to have invalid addresses we were still showing the 'Send' button.

So, we update the presenter logic to give us a property we can use in the view to detemine whether to show the button. We also amend the 'Ready to send' wording to highlight that the notice cannot be sent.

In the unit tests we do some housekeeping. We add some tests for properties not explicitly covered, and drop some that were no longer testing anything.