Refactor fetch returns recipients query
The initial query we developed for fetching recipients did not dedupe all the recipients.

We have developed a query that can now dedupe all the recipients.

This change updates the query to the improved query.

We have updated the tests to better align with these changes (and the previous tests were not so great, they would 'pick' the result elements rather than test the whole result). These tests do not check the result of the query, this has involved updating the 'LicenceDocumentHeaderSeeder'.

We now create the seeded data for each recipient type as and when we need it (we also added duplicate recipients so we can prove the update query is working).

We have not changed anything else. The existing 'DetermineRecipientsService' will remain for now but will be removed in another change.