Housekeeping - Fix going back to /manage
These should have been checking the feature flag to determine whether to go back to the legacy manage page or the system manage page.

As we have made the legacy one defunct with the removal of the feature flag, it's now not an issue. We just need to update all redirects to `/manage` to be `/system/manage`.