Highlight special char addresses (#2547)
https://eaflood.atlassian.net/browse/WATER-5341

A while back, we found an issue with our notice engine. It was sending addresses to [Notify](https://www.notifications.service.gov.uk/) incorrectly. In some cases, it was because we were passing too many address lines. In others, they were UK addresses with ‘country’ populated. So, the country was being added as the last line, but **Notify** requires the postcode to be the last line for UK addresses.

We fixed those issues, but whilst working on the changes, we also implemented a way of highlighting invalid addresses in the 'Check recipients' page. These were ones with no postcode or country, and therefore, there was no point in trying to send the notification.

As part of dynamic due dates, it's become even more critical that we endeavour to correct any invalid addresses (as per Notify rules) to improve the chance that notifications succeed the first time.

To help with this, we [added an invalid addresses report page](https://github.com/DEFRA/water-abstraction-system/pull/2531) to highlight these invalid addresses beforehand.

As part of this, we remembered that there is another reason Notify rejects addresses: when an address line starts with a special character. `@ ( ) = [ ] ” \ / , < >`.

This change extends the logic that highlights invalid addresses to include those that contain these special characters.