Undo `compareDates()` direct date comparisons
On reflection, using `compareDates()` to directly compare dates is overkill; it's far clearer to use `<`, `<=`, `>` and `>=` for date comparison and only use a helper function (`datesMatch()`) when checking equality, leaving `compareDates()` solely for use in sorting functions. We therefore revert the last few changes where `compareDates()` was used for direct comparison.