Delete invalid assertions
We realised that these assertions are checking what they think they are.

Without wading into detail, because we use `auth:` in the options passed to `server.inject()` we're telling hapi to authenticate without calling `validate()`.
```
server.inject({
  method: 'GET',
  url: '/some-protected-route',
  auth: {
    strategy: 'session',
    credentials: { userId: '12345' }
  }
})
```

The other version of this helper is to call `server.auth.test('session', { userId: '12345' })` before requesting the route.

But the TLDR is, our controller tests don't cause Hapi to call `validate()` in AuthPlugin, which means the `AuthService` never gets triggered, which means permissions never get set, which means the navbar is only ever populated with 'search'.

The only reason these currently work is because all our pages include the word 'management' in the footer!

It's a much bigger task to solve or refactor _actually_ authenticating in the tests in order to trigger AuthService. So, for now we just remove those assertions.