Drop rtn log isCurrent=true from send Rtn invites
https://eaflood.atlassian.net/browse/WATER-5343

When the Billing & Data team started checking the recipients that will be selected for the next round of return invitations they spotted there were some missing (when compared to a report they run).

When they looked, they were all linked to licences that had return versions with a reason of 'Succession or transfer of licence'.

When this is the reason against the return version, the return logs we generate from it have the flag `isCurrent`, found their their `metadata` field, set to 'false'.

We don't know why, it was something we just copied from the legacy code!

Another thing we copied was their query for selecting which return logs to base a returns invitation on. Their query excluded return logs where `isCurrent = false` so we did the same.

Now we're being told these should be included, and to drop this WHERE clause from our query. Hence this change.