Use end date not start date in return log download
https://eaflood.atlassian.net/browse/WATER-5285

When we built the new view return log page we added the ability to download it, specifically the return log's lines.

The version we built contains

- start date
- reading
- quantity

The business rightly pointed out this is not consistent with how we show return logs on the page. There the date we use is always the end date, especially when it comes to the lines as that is what the value is held against.

So, this change switches the date used in the download to be the line's end date.