Refactor date comparison (#2536)
https://github.com/DEFRA/water-abstraction-team/issues/141

This PR creates 2 new helper functions in `DatesLib` for date comparison:

- `sameDate()` takes 2 dates and returns `true` or `false` depending on whether or not the dates are the same;
- `compareDates()` takes 2 dates and returns `-1` if the first date is before the second; `0` if the dates are the same; and `1` if the first date is after the second.

`sameDate()`, as specified in the original issue, avoids the problem where dates have to be converted to numeric timestamps using `getTime()` before they can be checked for equality or inequality (otherwise, we are comparing whether the 2 dates are the exact same object); rather than litter the code with the same explanation multiple times saying why we aren't just comparing dates directly, we create the helper function and put the explanation there.

`compareDates()` is intended for use in sorting functions, where the return values of `-1`, `0` and `1` indicate the sort order.