Stop saving session data for too long (#2556)
https://eaflood.atlassian.net/browse/WATER-4992

This link will take you to the ticket for creating the view notice page. We've linked it because it was an issue with the filters saved in the session cookie that got us starting to think about our [Hapi Yar session](https://hapi.dev/module/yar).

In the first iteration, we did what we always do: on submit, we saved the filter values from the payload to the session.

```javascript
  yar.set('noticeFilter', {
    licence: payload.licence ?? null,
    recipient: payload.recipient ?? null,
    status: payload.status ?? null
  })
```

But then one of our team highlighted an issue. If you had previously been on the view notice page and used the filters, every time you came back to the page, the same filters would be applied until we cleared them. That's what happens on our other pages, but they are not specific to a record. For example, view notices you are always working with the entirety of the notice records.

Added to that, when you have finished sending a notice, we provide a link to view it on the confirmation page. If you click that, your filters from a previous visit would be automatically applied. But if they were set to `RETURNED`, for example, it would result in you seeing 'No notifications'.

Not what you want to see after just creating a notice!

So, for the view notice page, we [Stopped using a global filter session in view notice](https://github.com/DEFRA/water-abstraction-system/pull/2496), by which we mean we generate a notice-specific filter key.

But it got us thinking about our session. As it does not have a TTL, it defaults to 'session', which means it lives until the user closes the browser. Someone having a busy day viewing different notices is going to be adding a lot of these entries that could hang around for a very long time.

And thinking about it, if you added a filter in the morning on a page like view notices, you wouldn't expect the same filter to be applied to the page if you came back to it in the afternoon, when returning to the service.

The solution is to set a TTL on the cookie. But that means it will expire `X` milliseconds (we're going with 15 minutes to start with) after the first `yar.set()` call. We don't want that either.

So, along with setting the TTL, we need a mechanism to 'keep-alive' the session. For that, we add a new custom plugin that registers a function on [Hapi's onPreHandler step](https://hapi.dev/api/?v=21.4.3#request-lifecycle).

It [touches](https://hapi.dev/module/yar/api/?v=11.0.3#methods) the **Yar** session on every request, which in turn causes **Yar** to reset the TTL.

This means our 'session' and the data in it will remain alive whilst the user is using the service. After 15 minutes of inactivity, the session will reset. Hopefully, this will remove the feeling that filters become 'stuck'. But we'll continue to monitor the situation.