Fix return cycle seeding (#2562)
https://eaflood.atlassian.net/browse/WATER-5007

When we started work on moving returns management from NALD to WRLS, we needed to start working on return cycles.

Previously, [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import) handled creating them each year. But when we took over that responsibility that fell to us.

To support this, we added seeding return cycles to our seeder, so they would be available for unit testing.

In the main, everything is fine, and WRLS has been responsible for returns since June 2025. But every now and again, an environment will get two extra return cycles created: the summer and all year for the next year.

The problem is that when the team do anything with return versions, they'll see new return logs created for these future years, when we should only be creating them up to the current return cycle.

We've finally tracked down how this happens. It is when we run the seeder!

For example, imagine the current date is 21 October 2025. In this case, the _latest_ return cycles that should exist are

- 2025-04-01 to 2026-03-31 Winter and all year
- 2024-11-01 to 2025-10-31 Summer

The problem is, when the return cycles seeder runs is that it doesn't stop creating return cycles until its 'determination date' is greater than the _current date_.

The determination date starts off as 1959-04-01, and increments until it is greater than, in our example, 2025-10-21.

That means you end up with these additional return cycles.

- 2026-04-01 to 2027-03-31 Winter and all year
- 2025-11-01 to 2026-10-31 Summer

The loop used in the seeder needs to stop creating return cycles when the 'determination date' is greater than [current year]-04-01.

This change fixes the seeder.