Update return periods to include quarterly flag (#2565)
https://eaflood.atlassian.net/browse/WATER-5303

As part of our work to set due dates dynamically we can no longer rely on the due date when looking at return logs. This means when we're building a list of recipients to invite or remind about their due returns we need to use different criteria.

This PR updates the return periods to add the quarterly flag which will enable us to add it to queries.