Migrate monitoring station search (#2564)
https://eaflood.atlassian.net/browse/WATER-5354

We are migrating the internal search page (the root page of the internal service) from the legacy apps to our new `water-abstraction-system app`.

The existing page allows users to search using different types of values, for example, licence references, licence names, and return references.

We are implementing these different types of search into the new `system` page one at a time to manage the complexity.

Functionality to search for licences and returns has already been implemented in #2477 and #2551. This change is to implement searching for monitoring stations.

I still haven't added a feature flag for this feature yet as we can't really use it as a replacement for the existing search until we've migrated all the other things we need to search on. For the moment, this search page can just be accessed at `/system/search`.