Use UNKNOWN for notifications with no status
https://eaflood.atlassian.net/browse/WATER-5355

We've found approximately 8,000 water abstraction alert notifications that do not have a status. Nor do they have

- A Notify status
- A Notify ID
- Any information in their `log` field

We can only assume something went wrong when they were created that caused them to fail to record a notify ID, which then means the legacy service was unable to determine what their Notify status would be.

The issue is our new view notice and notification pages assume _all_ notifications have a status. When trying to display these ones we just get a tiny blue square.

We need to have a think about what to do with them. In the meantime we are going to apply a data fix to set their status to 'unknown' and ensure we can display them in the new pages.

We're not going to extend our filters for this status as we intend to determine a longer term solution for them.