Update fetch recipients query to not use due date (#2566)
https://eaflood.atlassian.net/browse/WATER-5303

As part of our work to set due dates dynamically we can no longer rely on the due date when looking at return logs. This means when we're building a list of recipients to invite or remind about their due returns we need to use different criteria.

This PR updates the query to use the end date, start date and also the quarterly flag. By adding the same options to the existing and new where clause with the only change being the due date to end date we can use the same bindings for both. This will make it easier to remove the block after we go live.