Meter readings not shown in download (#2569)
https://eaflood.atlassian.net/browse/WATER-5285

Following updates made to the return csv download to use the end date rather than the start date. The start date was subsequently removed from the fetch query in `app/services/return-logs/fetch-download-return-log.service.js` as it was assumed that it was no longer needed. Unfortunately this start date is needed for the return submission models `$applyReadings` method to work.

The fix is therefore to add the start date back into the fetch query. The unit tests will also be updated as they should have caught this issue.