Add return log ids to notifications
https://eaflood.atlassian.net/browse/WATER-5350

When we save a notification, we need to save the return logs ids.

When a recipient is associated with multiple return logs we will need to store them all.

This change updates the fetch returns recipients query to return any return log id's associated with the recipient.

We then save the return log ids when we save a notification.