Switch to non-batched simplified send notices
https://eaflood.atlassian.net/browse/WATER-5119

We recently had to track down why notices, specifically the email notifications linked to them, sent using our new engine in `production`, were not automatically updating their status.

We'd added checking the status as the last step of the batch sending process, so we could give feedback to users when the notice is sent, rather than having to wait till the next day.

We tracked down the issue to misconfiguration; basically, we were sending them all at once and not pausing long enough to allow [Notify](https://www.notifications.service.gov.uk/) a chance to send them.

But it meant digging into our existing 'batching' of the notifications for sending, and the additional steps it takes to avoid hitting [Notify's rate limits](https://docs.notifications.service.gov.uk/rest-api.html#rate-limits).

We're batching in an attempt to send the notice and its notifications quickly, and that adds additional complexity. But then we're having to add delays after each batch, currently of 10 seconds to avoid hitting Notify's rate limit.

We're also doing the same thing in our `ProcessNotificationStatusService`, which checks the the status of pending notifications.

Using our defaults, and assuming we have 1000 notifications to send the flow is

- send 125
- wait 10 seconds
- check the status of 125
- wait 10 seconds
- repeat

At that rate we could achieve processing 375 notifications a minute. This means it would take approximately 2.5 minutes to send the notice of 1000 notifications.

Now we're familiar with notifications, we also know this is rare. Generally users are sending a handful of notices containing a few notifications a day. The big 'sends' only happen twice a year.

So, lets got back to that rate limit. 3,000 requests per minute equates to 50 requests per second. If you sent them one at a time, you'd need to add only a 20 millisecond delay after each one to keep under that limit.

Up the delay to 30 seconds, and you get about 2,000 requests per minute. That's 1,000 below the limit, and in our example we'd be done in 30 seconds.

Add to that just sending them one after the other avoids the additional complexity of trying to batch the work.

So, this change simplifies our send notice and checks status processes to remove the batching.