Update alerts recipients to dedupe
https://eaflood.atlassian.net/browse/WATER-5350

We currently rely on the 'DetermineRecipientsService' to handle some of the deduplication when creating the recipient array.

This change updates the alert query to aggregate the licence refs for each recipient (where the contact hash type is unique).

This does differ from fetching return recipients, as we need to always return an additional contact if present (even with the same licence ref) and the primary use or licence holder. So we expect the additional contact and the primary user (or licence holder) to share a licence ref, this is only a concern when we need to merge the licence refs together.