Auto pause & retry when we hit Notify rate limit (#2578)
https://eaflood.atlassian.net/browse/WATER-5119

We recently had to track down why notices, specifically the email notifications linked to them, sent using our new engine in `production`, were not automatically updating their status.

We'd added checking the status as the last step of the batch sending process, so we could give feedback to users when the notice is sent, rather than having to wait till the next day.

We tracked down the issue to misconfiguration; basically, we were sending them all at once and not pausing long enough to allow [Notify](https://www.notifications.service.gov.uk/) a chance to send them.

But it meant digging into our existing 'batching' of the notifications for sending, and the additional steps it takes to avoid hitting [Notify's rate limits](https://docs.notifications.service.gov.uk/rest-api.html#rate-limits).

It got us wondering if we could simplify the whole process, whilst still ensuring we don't fall foul of the rate limit.

One step we can take is to automatically pause and retry if we do hit it!