Move Gotenberg delay into the request  (#2579)
https://eaflood.atlassian.net/browse/WATER-5119

We use [Gotenberg](https://gotenberg.dev/) to generate our PDFs, but the requests for them come once in a blue moon. This means we don't keep our Gotenberg instances running in ECS with lots of resources because it's a waste of money and energy (we are the EA after all!)

Because of this, when we do generate a PDF, we have found it is best not to overload our instances, so we use a slight delay between each request.

At the moment, the delay sits in `BatchNotificationsService`. But as it's a delay we would apply to _all_ requests, it should really sit in `GotenbergRequest` at the point we make it.