Remove start date from download recipients query (#2580)
https://eaflood.atlassian.net/browse/WATER-5303

We started filtering by start date in our `FetchDownloadRecipientsService` as part of updating our return invitation to stop using 'due date' as a filter (part of the new 'dynamic due dates' functionality).

That didn't go behind a feature flag because it wouldn't affect the results when we are using the old filter of `due_date = ?`.

We've just found it doesn't affect the results, but it sure affects the performance! On the big yearly invites of 3,500 recipients, it goes from 1 to 2 seconds up to 20 to 30 seconds! 😱

We need to go away and do some digging to find out why that is. But in the meantime, we need to whip out the start date and get the query performing again.